# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.RqChangePassword:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  auth.RqLogin:
    properties:
      email:
        type: string
      is_superadmin:
        type: boolean
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RqLogout:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RqUpdateUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  auth.RqUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      is_superadmin:
        type: boolean
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  auth.RsGoogleAuthURL:
    properties:
      url:
        type: string
    type: object
  auth.RsUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_superadmin:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  response.RsBase:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.RsError'
      message:
        type: string
      status:
        type: integer
    type: object
  response.RsError:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: Backend API for acareca
  title: Backend API
  version: 1.0.0
paths:
...
  /auth/google:
    get:
      description: get Google OAuth consent-screen URL
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RsGoogleAuthURL'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      summary: Get Google OAuth consent-screen URL
      tags:
      - auth
  /auth/google/callback:
    get:
      description: handle Google OAuth callback and return tokens
      parameters:
      - description: OAuth authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RsBase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      summary: Handle Google OAuth callback
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: login a user
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RqLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RsBase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      summary: Login a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: register a new user
      parameters:
      - description: Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RqUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.RsBase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      summary: Register a new user
      tags:
      - auth
  /auth/user:
    delete:
      description: Soft delete the currently authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RsBase'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Delete user account
      tags:
      - auth
  /auth/user/change-password:
    put:
      consumes:
      - application/json
      description: updates the password for the authenticated user
      parameters:
      - description: Password Change Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RqChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RsBase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Change user password
      tags:
      - auth
  /auth/user/logout:
    post:
      consumes:
      - application/json
      description: revoke the current session using the refresh token
      parameters:
      - description: Logout Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RqLogout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RsBase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Logout a user
      tags:
      - auth
  /auth/user/profile:
    put:
      consumes:
      - application/json
      description: Update the profile details of the user (email, names, phone)
      parameters:
      - description: Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RqUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RsUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.RsError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Update user profile
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    description: Type "Bearer <your_token>" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"